+++ AsyncLoggers/src/Wrappers/LogEventArgs/TimestampedLogEventArg.cs
@@ -1,10 +1,10 @@
 ﻿namespace AsyncLoggers.Wrappers.LogEventArgs;
 
 public class TimestampedLogEventArg(LogEventWrapper original)
-    : LogEventWrapper(original.Data, original.Level, original.Source, original.StackTrace)
+    : LogEventWrapper(original.Data, original.Level, original.Source, original.Timestamp, original.AppTimestamp, original.Uuid, original.StackTrace)
 {
     public override string ToString()
     {
-        return $"[{AsyncLoggers.GetLogTimestamp(this)}] {base.ToString()}";
+        return $"[{AppTimestamp}] {base.ToString()}";
     }
 }